refactor(hs): create shared directory in components
