+++ packages/headless-styles/src/components/Checkbox/shared.ts
@@ -2,7 +2,7 @@ import {
   createCheckboxFieldProps,
   getCheckboxFieldA11yProps,
   getDefaultCheckboxFieldOptions,
-} from '../sharedDefaultOptions'
+} from '../shared/defaultOptions'
 import type { CheckboxOptions } from './types'
 
 export function getDefaultCheckboxOptions(options?: CheckboxOptions) {
