+++ packages/headless-styles/src/components/Radio/shared.ts
@@ -1,4 +1,4 @@
-import { getDefaultCheckboxFieldOptions } from '../sharedDefaultOptions'
+import { getDefaultCheckboxFieldOptions } from '../shared/defaultOptions'
 import type { RadioOptions } from './types'
 
 export function getDefaultRadioOptions(options?: RadioOptions) {
